feat(ts/analyzer): Implement rules for some operators (#23)
stc_ts_errors:
 - Compact debug impl for `DebugContext`.

stc_ts_types:
 - Add accessibility to `Method`.

stc_ts_file_analyzer:
 - Add rule for assignment of consturctors in abstract context.
 - Allow `typeof M = M` where `M` is a module type.
 - Expand query types while calling.
 - Report an error if user tries to assign abstract class to non-abstract class.
 - Use better rule for assignment to empty abstract classes.
 - Error if the user tries to instantiate an abstract class.
 - Validate types of arguments.
 - Use non-dummy span while validating `TaggedTpl`.
 - Use generic-expanded type while validating argument types.
 - Handle `typeof s === 'string'`
 - Force type facts if required.
 - `infer_type`: Support `Interface` - `TypeLit`.
 - Handle `yield *` properly.
 - Use better logic for checking equality of keys while calling type elements.
 - Handle `any` in rhs in assigments.
 - Allow index signature eat multiple rhs while assigning.
 - Inference types from `IndexSignature` and `Property`.
 - Fix type of `number` + `number.`
 - Implement assignment of an enum to type elements.
 - Report errors for computed enum initializers within string enum.
 - Use intersection with module types while validating the object of a member expression.
 - try_assign_pat: Handle more patterns.
 - Add rules for binary operation (+).
 - bin: Handle undefined and null.
 - Check null and undefined in comparisons.
 - Better logic for checking overlaps in comparison.
 - Type literals can overlap.
 - Validate operands of comparisons.
 - Add rules for handling assignment of class methods.
 - Handle `foo instanceof any`.
 - Use `has_overlap` while validating comparison operands.
 - Check for null and undefined in lhs and rhs of `in`.
 - Expand refs while validating rhs of `in`.
 - Fix validation of lhs of `in`.
 - make_instance: Handle `null` and `undefined`.
 - Class and interface is can be lhs of instanceof.
 - Allow assigning `Function` to callable types.
 - Handle assignment to `Function`.
 - Support numeric comparisons, including implicitly casted types.
 - Report error for assignment of private class members.
 - Add some rule for validating arg of an update expression.
 - Validate arg of an update expr.

+++ typescript/file_analyzer/src/analyzer/assign/class.rs
@@ -7,6 +7,7 @@ use stc_ts_types::Class;
 use stc_ts_types::ClassMember;
 use stc_ts_types::Type;
 use swc_common::EqIgnoreSpan;
+use swc_ecma_ast::Accessibility;
 
 impl Analyzer<'_, '_> {
     pub(super) fn assign_to_class(&mut self, opts: AssignOpts, l: &Class, r: &Type) -> ValidationResult<()> {
@@ -14,33 +15,25 @@ impl Analyzer<'_, '_> {
 
         let r = r.normalize();
 
-        // Everything is assignable to empty classes, including classes with only
-        // constructors.
-        let is_empty = l
-            .body
-            .iter()
-            .find(|member| match member {
-                ClassMember::Constructor(_) => false,
-                _ => true,
-            })
-            .is_none();
-        if is_empty {
-            return Ok(());
-        }
-
         match r {
             Type::Class(r) => {
                 if l.eq_ignore_span(r) {
                     return Ok(());
                 }
 
-                // class Child extends Parent
-                // let c: Child;
-                // let p: Parent;
-                // `p = c` is valid
-                if let Some(parent) = &r.super_class {
-                    if self.assign_to_class(opts, l, &parent).is_ok() {
-                        return Ok(());
+                if !l.is_abstract && r.is_abstract {
+                    return Err(box Error::CannotAssignAbstractConstructorToNonAbstractConstructor { span: opts.span });
+                }
+
+                if !r.is_abstract {
+                    // class Child extends Parent
+                    // let c: Child;
+                    // let p: Parent;
+                    // `p = c` is valid
+                    if let Some(parent) = &r.super_class {
+                        if self.assign_to_class(opts, l, &parent).is_ok() {
+                            return Ok(());
+                        }
                     }
                 }
 
@@ -80,6 +73,20 @@ impl Analyzer<'_, '_> {
             _ => {}
         };
 
+        // Everything left is assignable to empty classes, including classes with only
+        // constructors.
+        let is_empty = l
+            .body
+            .iter()
+            .find(|member| match member {
+                ClassMember::Constructor(_) => false,
+                _ => true,
+            })
+            .is_none();
+        if !l.is_abstract && is_empty {
+            return Ok(());
+        }
+
         Err(box Error::Unimplemented {
             span: opts.span,
             msg: format!("Assignment of non-class object to class\n{:#?}", r),
@@ -92,16 +99,67 @@ impl Analyzer<'_, '_> {
         l: &ClassMember,
         r: &[ClassMember],
     ) -> ValidationResult<()> {
+        let span = opts.span;
+
         match l {
-            ClassMember::Constructor(_) => {}
-            ClassMember::Method(_) => {}
+            ClassMember::Constructor(lc) => {
+                for rm in r {
+                    match rm {
+                        ClassMember::Constructor(rc) => {
+                            self.assign_params(opts, &lc.params, &rc.params)?;
+                            // TODO: Validate parameters and etc..
+                            return Ok(());
+                        }
+                        _ => {}
+                    }
+                }
+            }
+            ClassMember::Method(lm) => {
+                if lm.accessibility == Some(Accessibility::Private) {
+                    return Err(box Error::PrivateMethodIsDifferent { span });
+                }
+
+                for rmember in r {
+                    match rmember {
+                        ClassMember::Constructor(_) => {}
+                        ClassMember::Method(rm) => {
+                            //
+                            if self.assign(&lm.key.ty(), &rm.key.ty(), opts.span).is_ok() {
+                                if rm.accessibility == Some(Accessibility::Private) {
+                                    return Err(box Error::PrivateMethodIsDifferent { span });
+                                }
+
+                                // TODO: Parameters.
+                                self.assign_with_opts(opts, &lm.ret_ty, &rm.ret_ty)
+                                    .context("tried to assign return type of a class method")?;
+
+                                return Ok(());
+                            }
+                        }
+                        ClassMember::Property(_) => {}
+                        ClassMember::IndexSignature(_) => {}
+                    }
+                }
+
+                if lm.is_optional {
+                    return Ok(());
+                }
+            }
             ClassMember::Property(lp) => {
+                if lp.accessibility == Some(Accessibility::Private) {
+                    return Err(box Error::PrivatePropertyIsDifferent { span });
+                }
+
                 for rm in r {
                     match rm {
                         ClassMember::Constructor(_) => {}
                         ClassMember::Method(_) => {}
                         ClassMember::Property(rp) => {
                             if self.assign(&lp.key.ty(), &rp.key.ty(), opts.span).is_ok() {
+                                if rp.accessibility == Some(Accessibility::Private) {
+                                    return Err(box Error::PrivatePropertyIsDifferent { span });
+                                }
+
                                 if let Some(lt) = &lp.value {
                                     if let Some(rt) = &rp.value {
                                         return self.assign(&lt, &rt, opts.span);
